backend/requirements.txt
alembic==1.16.4
bidict==0.23.1
blinker==1.9.0
boto3==1.39.13
botocore==1.39.13
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
eventlet==0.40.2
Flask==3.1.1
flask-cors==6.0.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-socketio==5.13.0
RapidFuzz==3.13.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
stripe==12.3.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
